assignment7c/milestone2/requirements.txt
django
django-filter
djangorestframework
django-rest-swagger
markupsafe
requests
